requirements.txt
Django==1.11.1
django-taggit==0.22.1
unidecode
psycopg2
django-tinymce
